Move 4C materials to four_c namespace
