+++ tests/test_cosserat_curve.py
@@ -32,7 +32,6 @@ import pyvista as pv
 import quaternion
 
 from meshpy.core.conf import mpy
-from meshpy.core.material import MaterialReissner
 from meshpy.core.rotation import Rotation
 from meshpy.cosserat_curve.cosserat_curve import CosseratCurve
 from meshpy.cosserat_curve.warping_along_cosserat_curve import (
@@ -42,6 +41,7 @@ from meshpy.cosserat_curve.warping_along_cosserat_curve import (
 )
 from meshpy.four_c.element_beam import Beam3rHerm2Line3
 from meshpy.four_c.input_file import InputFile
+from meshpy.four_c.material import MaterialReissner
 from meshpy.mesh_creation_functions.beam_basic_geometry import create_beam_mesh_helix
