+++ tests/test_four_c.py
@@ -33,7 +33,6 @@ import numpy as np
 
 from meshpy.core.conf import mpy
 from meshpy.core.geometry_set import GeometrySet
-from meshpy.core.material import MaterialReissner
 from meshpy.core.rotation import Rotation
 from meshpy.four_c.beam_potential import BeamPotential
 from meshpy.four_c.boundary_condition import BoundaryCondition
@@ -42,6 +41,7 @@ from meshpy.four_c.element_beam import Beam3rHerm2Line3
 from meshpy.four_c.function import Function
 from meshpy.four_c.input_file import InputFile
 from meshpy.four_c.locsys_condition import LocSysCondition
+from meshpy.four_c.material import MaterialReissner
 from meshpy.four_c.solid_shell_thickness_direction import (
     get_visualization_third_parameter_direction_hex8,
     set_solid_shell_thickness_direction,
