+++ tests/test_input_file_utils.py
@@ -27,10 +27,10 @@
 # SOFTWARE.
 """This script is used to test the functionality of the input file utils."""
 
-from meshpy.core.material import MaterialReissner
 from meshpy.core.mesh import Mesh
 from meshpy.core.mesh_utils import get_coupled_nodes_to_master_map
 from meshpy.four_c.element_beam import Beam3rLine2Line2
+from meshpy.four_c.material import MaterialReissner
 from meshpy.mesh_creation_functions.beam_basic_geometry import create_beam_mesh_line
